Updated CSVs
Not sure if these are the final CSVs that we are going with but they should have the proper amounts of data that we need for this project. I have already uploaded the Drinkers CSV into the AWS instance. So this version should be final. Setting aside a copy to make sure that we don't lose this