Frequents Update
Added in code in generator for Frequents table. Also created the CSV for frequents, there are 36,000 tuples. Even though we only need 5000 tuples TOTAL, we are hoping that the query pattern we implement will bring this down to 900 tuples. Our logic is that because we have 20 or so states, there is a 1/20 chance of the drinker being in state A, and a 1/20 chance for each bar to be in state A. This means that we expect a 1/400 chance for the drinker and the bar to be in the same state. We hope that are logic is correct because this will take forever to upload to our AWS instance.