Big Update for CSVs and Generator File
This would, ideally, be the final update that we need for the generator and the CSVs. We will be uploading the CSVs into the database for use. We may require further updates if the amount of tuples we generate do not match within a reasonable range of the desired 5000 tuples.